+++ src/components/layout/atoms/SliderItem/index.tsx
@@ -1,4 +1,4 @@
-import { Container, Img } from 'pages/StoreDetail/components/SliderItem.styled';
+import { Container, Img } from 'components/layout/atoms/SliderItem/styled';
 import type { VFC } from 'common/utils/types';
 import type { HTMLAttributes } from 'react';
