+++ src/components/layout/atoms/SliderItem/styled.ts
