+++ src/components/storeDetail/molecules/SeatLayout/styled.ts
@@ -1,0 +1,6 @@
+import styled from 'styled-components';
+
+export const Container = styled.div`
+  height: 38rem;
+  background-color: ${(props): string => props.theme.color.grey50};
+`;
