+++ src/components/storeDetail/templates/StoreDetail/styled.ts
@@ -1,0 +1,22 @@
+import styled from 'styled-components';
+
+export const BackBtn = styled.button`
+  position: fixed;
+  top: 4.4rem;
+  left: 1.6rem;
+`;
+export const HeaderWrap = styled.div`
+  padding: 2.4rem 1.6rem;
+`;
+
+export const Name = styled.h1`
+  font-size: 2.4rem;
+  font-weight: 700;
+
+  margin-bottom: 0.8rem;
+`;
+
+export const Introduction = styled.h2`
+  font-size: 1.6rem;
+  font-weight: 700;
+`;
