+++ src/components/tabs/atoms/TabNavItem/styled.ts
@@ -1,0 +1,10 @@
+import styled from 'styled-components';
+
+export const TabContainer = styled.ul`
+  display: flex;
+  height: 4.8rem;
+
+  box-shadow: inset 0px -0.1rem 0px #efefef;
+
+  /* background-color: yellow; */
+`;
